Refactor test packages <organisation
