+++ src/test/java/spaceinvaders/group_22/JavaArgumentReaderTest.java
@@ -10,6 +10,8 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
+import spaceinvaders.group_22.JavaArgumentReader;
+
 /**
  * Test for the argument reader.
  * @author Jochem
