+++ src/test/java/spaceinvaders/group_22/game/powerup/SpawnProtectionPowerupTest.java
@@ -1,4 +1,4 @@
-package spaceinvaders.group_22;
+package spaceinvaders.group_22.game.powerup;
 
 import org.junit.Rule;
