+++ src/test/java/spaceinvaders/group_22/logger/LogEventTest.java
@@ -3,6 +3,8 @@ package spaceinvaders.group_22.logger;
 
 import org.junit.Assert;
 import org.junit.Test;
+
+import spaceinvaders.group_22.logger.LogEvent;
 /**
  * Test class for spaceinvaders.group_22.logger.LogEvent.class.
  * @author Dorian
