+++ src/test/java/spaceinvaders/group_22/logger/LoggerTest.java
@@ -4,6 +4,9 @@ import java.util.ArrayList;
 
 import org.junit.Assert;
 import org.junit.Test;
+
+import spaceinvaders.group_22.logger.LogEvent;
+import spaceinvaders.group_22.logger.Logger;
 /**
  * Test class for the spaceinvaders.group_22.logger.Logger class.
  * @author Dorian
