+++ src/test/java/spaceinvaders/group_22/ui/UIElementBulletTest.java
@@ -4,6 +4,8 @@ import org.junit.Rule;
 
 import javafx.scene.canvas.GraphicsContext;
 import spaceinvaders.group_22.game.SinglePlayerGame;
+import spaceinvaders.group_22.ui.UIElement;
+import spaceinvaders.group_22.ui.UIElementBullet;
 
 /**
  * Unit tests for the GameUIController class.
