+++ src/test/java/spaceinvaders/group_22/unit/AlienBulletTest.java
@@ -1,5 +1,8 @@
 package spaceinvaders.group_22.unit;
 
+import spaceinvaders.group_22.unit.AlienBullet;
+import spaceinvaders.group_22.unit.Unit;
+
 /**
  * Test for AlienBullet, extends BulletTest.
  * @author Ege
