+++ src/test/java/spaceinvaders/group_22/unit/ShootPowerUpUnitTest.java
@@ -6,6 +6,8 @@ import org.junit.Test;
 
 import spaceinvaders.group_22.game.Player;
 import spaceinvaders.group_22.game.SinglePlayerGame;
+import spaceinvaders.group_22.unit.ShootPowerUpUnit;
+import spaceinvaders.group_22.unit.Unit;
 
 /**
  * Test for Shoot Power Up, extends PowerUpTest Test.
