+++ src/test/java/spaceinvaders/group_22/wave/DefaultAlienWaveFactoryTest.java
@@ -1,4 +1,4 @@
-package spaceinvaders.group_22;
+package spaceinvaders.group_22.wave;
 
 import org.junit.Rule;
 import org.junit.Test;
